taxeen/frontend1/requirements.txt
# Taxeen Frontend1 Requirements
# Flask + Jinja + HTMX + TailwindCSS

# Core Framework
flask==3.0.1
werkzeug==3.0.1
jinja2==3.1.3

# HTTP Client
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Security
itsdangerous==2.1.2
flask-login==0.6.3
flask-wtf==1.2.1

# Forms
wtforms==3.1.1
email-validator==2.1.0.post1

# Development
flask-debugtoolbar==0.14.1

# Testing
pytest==7.4.4
pytest-flask==1.3.0
